Disguises: Dash to Space
replace dashes with spaces in disguises.
This is necessary because channel setup commands are quoted meaning the quoted channel commands containing the disguise setup commands may not contain quotes themselves as these would clash with the quotes of the outer setup commands, as a consequence of this the disguise setup needs to happen without using any quotes which means that role names with spaces in their name cannot be used as a disguise. As a workaround for that I have been using hyphens to connect the two segments of the name. To return this into the proper name format, replace hyphens with spaces.